requirements.txt
mcp
newsapi-python
wikipedia
mcp[cli]
praisonaiagents
praisonaiagents[llm]